Massage customercare models to make tweet replies more traversible.
Tweet now uses tweet_id as its pkey. The surrogate key wasn't really useful, and this makes it simple to turn reply_to into a Django ForeignKey so we can traverse it for the Answered and Unanswered filters. Cascading deletes are okay for replies; it doesn't make sense (UI or otherwise) to hang onto replies to tweets we no longer have.