Create a PostCrash API for Socorro. [bug 631341]
The API is available at /postcrash. There is one required GET parameter: s.

A request without an `s` parameter is a 400.
A request with an `s` parameter that does not return a result is a 404.
A request with an `s` parameter that returns a result is a 200. (Or 304.)

All responses are cached for 24 hours and have Content-type: text/plain.

If an article matching a crash signature is found, its URL isreturned without
a locale. This will cause SUMO to use our Accept-Language localization
fallback when a user visits the URL.