Ensure label and input text baselines match [bug 647026].
I lined up the labels with their targets such that the baseline of the
label and the baseline of the text in the input field line up.

Only tested on Windows. Please test on Mac. (Pixel-perfection cross platform
is not expected.) Please test at default zoom level: other zoom levels are
an at-your-own-risk deal.