Add ReverseProxyMiddleware to support running in a separate process.
Django's HttpRequest.is_secure() depends on an environment variable, which
is great when we're running with mod_wsgi inside the Apache process and can
rely on SetEnv HTTPS on, but when we're running as an isolated process (via
gunicorn or uWSGI or what have you) that doesn't work.

One of the upstream proxy servers (maybe even Zeus) will set an X-Scheme
header to 'https' or 'http'. With this middleware, we check the header in
is_secure() instead of relying on the environment variable.

I'm leaving this to be installed in settings_local for now, since this is
only for a test.