Cache total number of questions most of the time. [bug 661547]
Cache the count of visible questions for 15 minutes for most views. Not
caching for "my contributions" or tag filtering right now. Will want
to add both eventually, though maybe with shorter TTL.

Since only the filter affects the total, the cache key is based on the
filter only. It speeds up subsequent page loads, or switching the sort,
significantly.

Also, we were stomping on the built-in "filter", so I fixed that.