[bug 662749] Add a notification for when revs are ready for l10n. Redo subscription UI.
If someone subscribes to both Ready revisions and Approved ones and a new
revision is approved with the Ready box checked, they receive only a Ready
mail. If someone subscribes only to Approved revisions, they get only an
Approved mail.

Upgrade to django-tidings 0.2 to make the above possible.

As for the UI, instead of 4 hard-to-read links in the right column of the l10n
and contributor dashboards, we now do a jquery.selectbox-like popup with a
matrix of checkboxes. This also avoids reloading the page and redirecting to
possibly the wrong dashboard.