[bug 667892] Don't count translations whose parent docs lack any ready-for-l10n revisions when computing the numerators for the Overview readout.
Also take the unnecessary stuff out of the TemplateTranslationsReadout query, and remove the uninformative Visits column from the display. This fixes [bug 643427]. In optimizing that, the Templates readout grew to have little in common with the MostVisitedDefaultLanguageReadout, so I broke that inheritance relationship and factored out the common stuff to a top-level _format_row_with_out_of_dateness().