First pass at moving sphinx stuff to SphinxMeta
This is a first pass at moving some of the stuff in search/client.py to
SphinxMeta classes on the models in question.  This doesn't actually
remove the code in search/client.py, though, and this doesn't affect
the behavior of Kitsune search at all--it's just a first pass at paving
the way.