[bug 697715] Don't delete Forums or Threads when the user who made their last post is deleted.
For now, we just set last_post to NULL. Django doesn't bring us what we need to set it properly, but it might be possible in the future with a custom on_delete value.

Stopped saying "No posts" when there's a null last_post, since it can also mean "We don't know what the last post was" now.

Added model makers for forums, because I wanted to use them.

To do:
* View tests to prove I fixed all the non-null assumers
* Porting all this to kbforums as well
* Auditing the rest of the ForeignKeys in the codebase