Rework results handling to be more efficient
Previously, we did a db hit for every search result in each group
(wiki, forums, questions) that was being queried.  Now, we only do a
db hit for each search result that's being displayed and further, we
now do the queries in batches.

This works by fiddling with lists of object ids to figure out exactly
what we're showing and then slicing the groups such that we're only
looking at search results that are going to get shown.