Remove deepcopy calls
This is actually probably a bad idea.  The problem here is that now that
we're using settings_test and that test_utils copies _everything_ that's
seen in dir() which includes built-ins and everything else, is that
deepcopy inevitably bumps into something that can't be deepcopied.

For now I've switched deepcopy calls to copy.  Academically, this is
the wrong thing to do, though, since it's a blind fix.  What should really
happen is that tests that change settings should be fixed to change
settings back even when the test fails in the middle.  Might be easily
doable by mocking the settings module.  Practically, the tests pass and
everything seems ok, so it's probably the case that either settings
are shallow-enough or tests aren't futzing with things deeply.