Rework question indexing
Instead of having a document in the index for every Question AND Answer, we
now have a single document for a Question and all its answers all packaged
up and tied with a ribbon.

The one thing that changes slightly data-structure wise is the answer_creator
field.  This changes it to be a list of strings.  I'm not entirely sure if
that'll work, but we can make it work later.  Maybe with a big enough
hammer.