Change elastic indexing of forum threads
In elastic we now base the index on Thread and not Post.  That causes
problems with Oedipus, so we do some fancy waffling to deal with that.

Additionally, in elastic, we index a single Thread and all its Posts
as a single document rather than as separate documents.  This change
is to work around not being able to do a group_by like thing in elastic.