Fix failing tests
Quoting the comment I modified and added to:

If we're doing elasticsearch, then we need to update
the _s variables to point to the sliced versions of
S so that, when we iterate over them in the
following list comp, we hang onto the version that
does the query, so we can call excerpt() on it
later.

We only need to do this with elasticsearch.  For Sphinx,
search_s at this point is an ObjectResults and not an S
because we've already acquired object_ids on it.  Thus
if we update the _s variables, we'd be pointing to the
ObjectResults and not the S and then excerpting breaks.