[bug 709202] Wrote test for issue
Wrote a test that fails because of the issue.  Ipso facto, when this test
passes, I have fixed the issue.

In doing this, I moved the LocalizingClient up to the class and added an
answervote marked as helpful to the fixtures.  In doing the latter, I had
to tweak the expected results for a test because it now picks up a result
it didn't pick up previously.