[bug 715536] Multi-index-task squashing
Incremental updatie (not delete!) tasks all get tossed in a thread-local
queue. At request_finished, there's a handler that creates tasks from the
queued tasks. Because the queue is a set, this removes task duplicates
so we're only creating one index task per request per document.

This also fixes refresh calls so they pass timesleep=0--we don't need to
wait for a second to update the index.