[bug 718456] fixes approved mail
When approving a revision, kitsune sends email to people subscribed to
approvals. The email now has the approver of the revision (not the
creator of the revision) in the email body. The email also now has
the diff of the newly approved version from the most recent approved
version.

In the process of doing that, I:

* removed a couple of carriage returns from the email template
* updated the test code
* cosmetically changed some code removing end of line \s
* wrote a big TODO in the test code about how I think the test
  should get fixed so that it's less duplicative