[bug 723718] fix date and string indexing
We index and query dates in Sphinx as an int. Prior to this commit, we were
indexing dates in ES as a date which doesn't compare correctly when we
query with an int. So that's fixed now.

Additionally, we were running analysis on tags and username fields--we
shouldn't do that since these should be matching exactly.