[bug 706948] Fix num_answers mismatch
self.question.answers.count() was pulling from cache, so when we create
a new Answer, it wasn't reflected in the count and then question.num_answers
would be 0 which was a mismatch.

This forces it to come from the db and not cache and thus we get the
right answer.