[bug 739794] Fix index prefix
So... -dev and -stage share an ES cluster and the problem that creates
is that it makes it impossible for us to test production deployments
on -stage without also hosing -dev.

What this does is changes the nature of ES_INDEX_PREFIX. Instead of
using it to filter out which indexes are pertinent to SUMO on a given
server, we now prepend the prefix to the index name. This allows us
to have two servers with different prefixes with the same ES_INDEXES
and ES_WRITE_INDEXES and the resulting index names will be different.
It also allows us to change the ES_INDEXES and ES_WRITE_INDEXES for
a deployment that requires a reindexing and everything should work
on both -stage and -prod.

This also moves the Sphilastic code from oedipus to kitsune because
I was sick of changing it in oedipus and then updating kitsune with the
new commit.

This also includes a new version of elasticutils that allows us to change
the way the index name is calculated for an S.

Also, I did some paragraph wrapping by accident and that's part of this
commit, too.

As a side note, instead of fiddling with this and working around the fact
that elasticutils wants to be clever about index names, I wish I had reworked
elasticutils so it was less clever and more useful. But that is a story for
another time.