[bug 745473] Fix typo
pyes switched from indices to indexes (or vice versa--I can't remember).
Point being that we were using the wrong one the result of which being
that it used the default index created with the ES object which calculates
the index name differently. Ugh.

This fixes that by changing the argument from indices to indexes. Now it
uses the one calculated in get_documents and all is well.

This also adds a test for get_documents.

And while doing that, I tidied up the import statements.