[741801] Create test_sphinx, move tests over
I moved a bunch of tests over to test_sphinx which still uses fixtures,
but when we're done with Sphinx we delete the file and we're golden.

There were a handful of tests that tested the search view and were
search system agnostic. So I kept these in test_search and tweaked
them so they didn't use fixtures.

So now we have test_search.py which is search system agmostic,
test_sphinx.py which has Sphinx-specific things, and test_es.py which
has ES-specific things.