[741801] Fix Sphinx and ES tests to skip
If sphinx isn't installed or configured, then this will skip tests that
require it.

Previously, it would only skip tests if SPHINX_SEARCHD or SPHINX_INDEXER
weren't defined in settings, but since they are defined in settings.py, that
was never the case unless you explicitly undefine them in settings_test.py.

Now it also looks to see if the set paths exist and if not, it skips them.

This fixes the ES test case, too. Moves the ElasticTestCase class it from
the sumo app to the search app and fixes it so it checks to see if ES is
running and if not, skips ES related tests.