[bug 746609] Rewrite ES index code unify doctype
This is kind of intense. It reworks all the indexing and searching
code so that we're using a single ES doctype in the index. Documents
can be distinguished by their 'model' field which indicates which
model they derive from.

This pulls in a change to elasticutils that adds a get_doctype()
to S. Previously the doctype was based on cls._meta.db_table and that
made a unified doctype difficult.

I added a .search() method to SearchMixin. This is the class method
to use to get an S that is tuned for that model.

I changed a bunch of tests to use the new .search() rather than building
their own S or Sphilastic.

Additionally, this does some minor cleanup of related code.

One thing this doesn't do is rework how the AAQ suggestions work. That's
still a single function that does its thing. Need to keep an eye on that
when we make future changes so we make sure it doesn't break.