[bug 706948] Call update on queryset instead of save on instance.
We are passing a question to task that might not run for a while
and then calling save on that instance. This may lead to stale
data overwriting newer data. This may explain the increase in
incidence of the no-replies issues recently, as celery seems
to be getting backlogged more frequently.