[bug 706948] Pass question id instead of instance to task.
We were passing a question to a celery task that might not
run for a while and then calling save on that instance.
This may lead to stale data overwriting newer data. It might
explain the increase in incidence of the no-replies issues
recently, as celery seems to be getting backlogged more
frequently.