[bug 752519, 752517] Filter and Boost ES fixes
Prefixes all the fields making it easier to apply boosts on a model-field
level and makes it less likely filters will interfere with each other.

Reimplement query_fields so that the code is in kitsune rather than in
elasticutils

Split _search_suggestions in the new_question view so that we can do
ES-specific things in the ES version of the view.

Add some AAQ tests that make sure the new_question view isn't totally
dying in an unhandled way.

Removed weights because they were using the wrong keys which means that
they weren't having any effect. The weight keys were wrong. e.g. a weight
specified "title" but since we're querying with "title__text" it needed to
specify "title__text".

I also updated to an elasticutils sumo branch rev that has some tweaks
to the weight code which makes it clearer.

Also added some debugging.

Fixed the result template to be more generic making it easier for our
increasingly divergent ES and Sphinx code.

Fixes the model field so it's not stored in the index--that was silly.
"store" should only be used for data that will be highlighted. Otherwise
we pull it from the source.

Removed question_id and thread_id from the mapping--they were never used.

Whew!