Prefix all the things!
But seriously, we decided to prefix everything because it makes it far
more obvious where data is coming from when looking at it downstream.

Plus prefixing is kind of like a poor-man's namespacing and that'll make
it easier to boost things for specific models rather than for everything.

Fixes the result template to be more generic making it easier for our
increasingly divergent ES and Sphinx code.

Also fixes the admin index page to deal with the newly prefixed title
fields.