[bug 757666] Add a prefix to _id
This changes _id (the document id) to include the model name. That way if
two documents have the same db id, the last-to-be-indexed document doesn't
stomp on the prior ones.

Note that _id is the elasticsearch document id and shouldn't be confused
with document['id'] which is just a field in the document named id which
matches the id in the database.

Also note that at the moment, elasticutils assumes that _id is an int, so
doing anything that causes it to do ObjectResults kicks up an error. That
needs to get fixed, but can be pushed off until I figure out how to do it
in a way that's not fundamental to how elasticutils thinks about the world.