[bug 762018] Fail if no index and no --delete
If you run:

    ./manage.py esreindex

and the ES_WRITE_INDEXES doesn't exist, then this will now fail and tell
you to use the --delete flag.

Why? Well, because if the index doesn't exist, it needs to get created
with the mappings. I figured it's better to just fail in this case rather
than go and do stuff and maybe you don't notice what happened.