[bug 762018] Fix reindexing if there's no index
If you run:

    ./manage.py esreindex

and the ES_WRITE_INDEXES doesn't exist, then this will now fail and
tell you to use the --delete flag.

Why? Well, because if the index doesn't exist, it needs to get created
with the mappings. I figured it's better to just fail in this case
rather than go and do stuff and maybe you don't notice what happened.

With the admin, if the index doesn't exist, we can't reindex into it.
So nix the option from the admin page altogether.