[bug 765376] Fix AS ES questions sortby
That's Advanced Search questions sortby in ES. The problem is that
ES index doesn't have a "replies" field--it's called
"question_num_answers". So if someone used this sort then it would kick
up an error and die.

As a side note, this has been broken for months and no one mentioned
it. So maybe this is something we can remove since no one uses it?