Fix unified search filters
If a search adds filters on model Foo, but not on other models, then
what happens is that the Fs get combined and the resulting F requires
things that are only in model Foo and thus the search results only
have results from model Foo in them.

This fixes that by forcing the F to specify which model they filter on.
So when they get combined, we get filters like "if the model is Foo AND
this AND that" OR "if the model is Bar AND this AND that" ... which
alleviates the combination problem.