[bug 765207] Fix update_question_vote_chunk
This task would do some ES index updating. The problem is when ES was
down, it'd send a flurry of emails to us plus it wouldn't update
the index.

This fixes that. If ES is down, it'll toss all the questions that
need to be updated into an index_task which retries when it fails.
Now we won't get more emails and the question data won't get
stale.