Clean up test_es tests
This cleans up the tests. Setting up the data had a lot of extra goo
in the code, so it wasn't easy to understand what was actually needed
by the test and what wasn't. Additionally, I ditched a lot of "color"
text.

I removed one test that isn't needed anymore now that we have a single
SORT_QUESTIONS set.

I nixed a bunch of lines by adding save=True to the function calls.