[bug 774761] Use unicode when truncating JSON.
Before, when JSON was truncated that contained unicode, the unicode
would be represented as escape characters when calculating how much to
truncate. The string to be truncating though would still be represented
as a unicode string, so

Example: The dictionary {'pi': 'π'} would be serialized as '{"pi":
"\u03c0"}', which 16 characters, instead of 11 if it were represented as
unicode. The truncation happens on the original string, which has not
been encoded as escape characters. If the limit is 12 characters, then
the truncation goal would be to chop of 4 characters, which are not
available (since it is only a single unicode character), causing an
error to be thrown.

This commit add the parameter ``ensure_ascii`` to turn this behavior on
or off. It is off by default.