[bug 768226] Implement m2m product field
This adds a products field to wiki Document which is a many-to-many
with the Products table.

It was tricky because we used "products" in the form and then had
some code to move things to tags. I had to rename that form field
to "product_tags" in order to fix a bunch of issues that adding
the model field created. Then I had to carry the product_tags form
field name change through the code base.