[Bug 701385] "Update" questions only when answered.
Previously questions would have the `updated` field set to
`datetime.now()` when ever `Question.save()` was called, unless
`no_update=True` was specified. This caused it to move to the top of
question listings, which is not desired. The only time a question should
be moved to the top is when it is answered, so this changes the default
behvior to *not* update the `updated` field unless `save` recieves a
parameter `update=True`. The `Answer` model has been similarly modified.

Tests have been added/modified to assure that this is the case, and I
looked over existing code to make sure anything that interacted with
`Question.updated` was adapted.