[bug 771716] Remove bucketed search view
This removes all the bucketed search view code and most of the code it
required. There are some straggling pieces, but we can remove those
at a later date and under other bug numbers.

This removes the esunified waffle flag and the e= url parameter in all
the search urls.