Use the difflib, Luke.
This reworks the troubleshooting page to (ab)use difflib and show us
an easier to parse page of indexed items vs. items recently saved in
the db.

This is all admin stuff, so if it doesn't work, it has no impact except
that I'll probably commit a fix in my leisure so I can look at the
troubleshooting page again.