Fix logarithmic retry code
I bumped into this while fixing something else. The retry code used
to have an off-by-one and I must have fixed it in a way that it's now
essentially an off-by-two.

The gist of it is that prior to this fix, it wouldn't go through all
the retry attempts--it'd skip the last one. With this fix it now
goes through all the retry attempts.

Also, I fixed some syntax since we want to use the:

    except Exception as exc:

syntax now.