[bug 778771] Use uncached kb data for index
Scoobidiver updated the keywords for a kb document on August 14th.
The new revision was saved and the index was updated. However, the
document in the index doesn't have the same keywords data.

1. I wrote a test (added in this commit) which simulates that and it
   runs fine, so I suspect there's an out-of-our-control thing going
   on.

2. Scoobidiver told me about two such documents, but I was only able
   to verify one of them. I have no idea how often this happens.

3. My current wild guess is that it's a caching issue. So I tweaked
   extract_document to pull from the regular ObjectManager rather
   than the caching one. It's a wild guess--no clue if it'll help
   or hinder.

Other than Scoobidiver telling us when there are problems, we don't
have a good way of measuring the problem or even seeing other instances
of it. If the problem continues, that's the next step.