Remove LOG_LEVEL settings.
This setting is misleading, because overwriting it in local_settings.py
would not have the desired effect. Since the dict is built at import
time, local_settings.py would have no chance to modify it before the
dict was built, and so nothign would be changed in the log settings.