[bug 790425] A/B test excerpt styles
This adds a search-ab waffle flag for future A/B testing of search
behavior.

I use it in this case to A/B test the excerpt style for support forum
questions.

* True/a shows the first 500 characters of the question_content.
* False/b shows 3 fragments of question_content as selected by ES.

After we're done A/B testing this, we can leave the search-ab stuff
in for the next round of A/B testing goodness.