Expand Special case for showing windows on mobile.
When the default mobile browser is selected, the default desktop browser
is also shown. This means that if you pick Android, Windows 7 will be
shown as well. However, {for="win"} sections should *also* be shown.
This adds a special case to handle treating "win" as the same as Windows 7
for this case.