[bug 682730] Add ability to edit a translation w/o making it up to date.
This adds a checkbox to the submit new l10n revision modal. When it is
checked, the new revision keeps the based_on of the previous revistion
being edited instead of updating it to the latest approved and ready for
l10n en-US revision.

This allows edits on translations to be made without marking them as
up to date.