[Bug 809433] Add registration surveygizmo.
In doing this, I noticed that the other surveygizmo on the site was
powered by waffle, and had a comment saying that it should probably be
in another file. I moved the old surveygizmo, and this new on, to an
external file controlled by a data attribute on <body>.

Also added a migration (177) to add the waffle flag used.