[Bug 809092] Karma badges only consider 3 months
Change the default daterange for karma badge calculations to be only 3
months, instead of all time. This was already parameterized, I just
changed the default and updated the tests.

This commit also makes daterange a required paramater for all karma
manager methods (except `count()`), instead of being optional and
defaulting to 'all'.