[Bug 801587] Add download button to Firefox page.
Add a download button styled after the firefox download pages found
elsewhere to the Firefox product page. This doesn't direct link to a
download, because that would duplicate a lot of logic from elsewhere.
Instead it links to a localized download page that offers the correct
download for locale, system, and version.