[Bug 808151] Send review comments to reviewer too.
This adds the reviewer to the list of recipients when a notification
gets sent for a revision being reviewed. This might be confusing for the
reviewer, as the email is targeted at the person who wrote the revision.
It will also reveal the email address of both parties to eachother, and
allow an email conversation to happen. All of this might be undesired,
but is what was requested.