[Bug 808151] Send review comments to reviewer too.
This sends two emails when a review goes out. One for the reviewer, and
one for the person who wrote the review. They are independent emails, so
this doesn't reveal email addresses, but also doesn't allow out-of-band
communication. This might also be confusing for the reviewer, if they
aren't expecting the email.