[Bug 807470] Make cards fit cards on product page.
This removes the sidebar, like suggested in the bug, but even with the
style changes Bram suggested, long languages like German still made the
cards uneven., I added a bit of JS to expand the cards to the same
height as the largest in the group. This is noticeable on page load, but
it looks just like reflow due to image loading in most cases, so I think
it is acceptable.