[Bug 820331] Fix registration.
This broke because the action of the registration form was the new auth
form, which does not understand POSTs, instead of the registration view,
which does.

This regression was caught by QA, and really only makes sense to test i
Selenium or similar, which is why I haven't also added a test for it.