[Bug 815670] Add buttons to go to mobile site.
On a small device that is viewing the desktop site, a dismissable banner
will appear at the top of the site, prompting the user to go to the
mobile site. If the user just moved from a mobile page to a non mobile
page, the banner will all the user to go back.

This also adds a middleware that allows us to force a client into either
mobile or desktop view by adding `?mobile=0` or `?mobile=1`, which is
how the "switch to mobile" button works. This is useful for testing.

On large screen devices, these banners are hidden with media queries.

Feedback