[Bug 821274] Generate AoA top contributors stats.
We used to use Mozilla Metrics to measure this. We stopped showing it in
the new design, but re-added it recently. Unfortunatly, it has appears
that Mozilla Metrics is no longer updated.

This commit removes all the old Mozilla Metrics code. It also removes
the code relating to statistics about overall tweet activity, because we
don't currently monitor that (and it wasn't shown anyways).  To replace
the removed code, I added code to generate statistics from the data we
are already collecting about replies that go through the AoA system.

Also made some changes to the tests surrounding these features, and the
format the data is stored in, since the view was previously transforming
the data into this form, so might as well do it in the cron job.

- Deleted some now unused JS.
- These stats now run every hour via cron, instead of six hours.
- Add a migration to update the data in redis.