[Bug 826277] Set DJANGO_SETTINGS_MODULE in deploys
Freddo, which manages our -dev deploys, is a Django app. When Freddo is
run, it has the environment variable DJANGO_SETTINGS_MODULE set to
accomodate Freddo. Unfortunatly, Kitsune tries to use this value as
well, which messes things up. Kitsune tries to import 'freddo.settings'
when getting Django settings, which fails.

This change resets the environment variable to 'settings_local', which
exists on the kitsune path, and so things succeed.