Add Products and Topics FKs too.
Topics and Products are M2M right now mainly because adding FKs is so
hard. We aren't using this in the code yet, but if we are going to go
through the process of doing an alter table, we might as well add these
anyways.

TODO: Eventually convert the code to use this too.