[bug 825031] Change chunk size
This drops the chunk size from 50k to 20k. This will increase
the number of indexing chunks that get created and spread the load
across more workers. This will increase the load on ES (more things
being indexed at the same time), but probably shouldn't affect
mysql at all.

As a side note, since we set the chunk size at 50k, mysql got some
upgrades and shifts and is better now and we updated ES to 0.19.11
and it's doing better now, too.