[Bug 813472] Use addon for about:supprt data.
About:support data is historically rarely actually submitted, which
makes things hard. This change prompts the user to install an addon
which will automatically fill out this important field for the user.

Currently this has a few problems revolving around the addon not being
perfect and being hosted on people.mozilla.org instead of somewhere
better.