[Bug 813472] Use addon for about:support data.
About:support data is historically rarely actually submitted, which
makes things hard. This change prompts the user to install an addon
which will automatically fill out this important field for the user.

After the addon is installed, it triggers an event, and the page fills
in the appropriate data. If the addon is already installed at page load,
it will load the data immediately.

The addon only provides this data to https://support.mozilla.org/ (NB:
SSL). To manually test this, you can launch Firefox with an environment
variable set to "bless" other urls. For example

    TROUBLESHOOTER_ORIGIN="*" firefox

Will cause all pages to have access to the troubleshooting information.
You can also give a space seperated list of urls, such as
"http://localhost:8080 http://support-dev.allizom.org".