Add xx locale for testing
This adds:

1. xx locale which is Pirate
2. scripts/poxx.py which translates a .po file into Pirate
3. scripts/test_locale.sh which extracts strings and runs poxx.py

The xx locale is set up so that it doesn't show up in the language
chooser lists, but you can explicitly specify it in the url:

    http://localhost:8000/xx/home

The locale is only available translated locally--it will show up
as English on -dev, -stage, and -prod. It'll show up as English
if you don't run scripts/test_locale.sh.