[bug 839448] Fix resend confirmation email edge case
If the user registers, then gets an activation email, then clicks
on the link thus activating the account, then the corresponding
RegistrationProfile is cleaned up, then the user clicks on the
link again, the code would traverse a path where it tries to
render a Jinja2 email template using the Django template renderer.

This fixes that. Plus a test case.