[bug 845165] Allow email changes if you're not logged in
The template errored out if you weren't logged in, but when you're
changing your email address, you request the change, then wait for
the email with the link, then click on the link--there's no guarantee
that the user is logged in.

This nixes the crumbs if the user is not logged in thus alleviating
the error.