[Bug 844933] Only use well formed troubleshooting
Instead of trying to massage and carefully tip toe aroud the parsed
troubleshooting data, this puts a filter on it so only something that
has 100% of the fields that the template uses and has all the expected
types is used. When the data doesn't match these expectations, it will
appear as plain text, unformatted. This errs on the side of not being
friendly enough, but being safer.

Hopefully.