[Bug 847874] Plot topic activity over time.
* Updates elasticutils to add support for histograms.
* Shows volume of questions, broken down by topic.
* Pulls data from ES.
* Configurable date range and bucket size.
* Can reorder and hide topics on the graph.

This commit adds a customized version of jQuery UI. This version of
jQuery UI has the Sortable and Datepicker features. It also has a small
tweak because datepickers were erroring out.

We already have jQuery UI, yes, but it is customized to be a subset.
Adding things after the fact is a PITA, so @rdalal suggested I just make
a new bundle that had what I need in it. It gets bundled with
jingo-minify anyways.