[bug 838587] Add support for youtube embeds.
This overloads the [[Video:]] wiki tag so that if the you pass
a youtube URL, a youtube embed is rendered to the article.

We need to do this in two steps because we don't allow iframes
in the markup. So we do a first pass by putting in placeholders,
replace the placeholders at the end with the real embeds.