Add a customized version of jQuery UI.
The version of jQuery UI has the Sortable and Datepicker features.

We already have jQuery UI, yes, but it is customized to be a subset.
Adding things after the fact is a PITA, so @rdalal suggested I just make
a new bundle that had what I need in it. It gets bundled with
jingo-minify anyways.