[Bug 785850] Limit votes on answers and questions
Limit votes on answers and questions to 10 votes per day of either type.
Votes after the limit of 10 will be silently ignored. The limit is based
on IP addresses for unauthenticated users, or user id for authenticated
users. Authenticatd users are not limited by IP at all.

User IP address is based on X_CLUSTER_CLIENT_IP, if available, or
REMOTE_ADDR if not. X_CLUSTER_CLIENT_IP has shown to be harder to spoof
than X_FORWARDED_FOR.