[bug 831005] Update to elasticutils master tip
* add document_id to index
* update to elasticutils master tip -- this picks up pyelasticsearch
  and ditches pyes
* nix get_indexing_es (it was dumb)
* remove bunch of now-unused settings (also dumb)
* tweak test scaffolding -- this should make tests run faster
  for kitsune in a VM and on slower machines
* clean up docs

Note: This doesn't remove pyes. For the purposes of making this easier
to review, that'll get done in a future commit.

Note: This adds a new field to the mapping, but it's only used in indexing
and it's obviously a string. Thus it's ok for ES to infer what it is
and we don't have to do the "change the mapping" two-step dance.