[Bug 857153] Add data bucketing to rickshaw_utils.
This gives the ability to reduce time regions in buckets to make things
a little cleaner to analyze on the large scale. The reduction method
defaults to simple sum, but can be configured to average for things like
percentage lines.

There are also tweaks to the set and line system to make this all work
together, and there is a very useful update method on Graphs now which
redraws the graph in a pretty reliable way.