[bug 838638] Add SearchMappingTypes and rework indexing (1/2)
This adds a parallel universe of indexing which the es_reindex_cmd
and admin code will use to create the new index with the new
SearchMappingType code.

It also adds a bunch of TODO lines for things that need to be fixed
in the subsequent commit.

Comments out one test that gets whiny with the parallel indexing
worlds.